refactor: move diabetes/frontend para frontend/diabetes
