+++ src/app/lista-nominal/page.tsx
@@ -1,3 +1,3 @@
-import { AcfPage } from "@/features/acf/diabetes/frontend/AcfPage";
+import { AcfPage } from "@/features/acf/frontend/diabetes/AcfPage";
 
 export default AcfPage;
