+++ src/features/acf/frontend/diabetes/AcfPage/modules/PanelSelector/consts.tsx
