+++ src/features/acf/frontend/diabetes/AcfPage/modules/PanelSelector/modules/List/modules/CoeqDataTable/modules/CpfOrBirthdayFormatter.tsx
