+++ src/features/acf/frontend/diabetes/AcfPage/modules/PanelSelector/modules/List/modules/CoeqInternalcards/presentation.tsx
