+++ src/features/acf/frontend/diabetes/AcfPage/modules/PanelSelector/modules/List/modules/print/getPrintDataResponse.ts
